AI_SERVER/requirements.txt
Flask==3.0.3
pandas==2.2.2
numpy==2.1.1
scikit-learn==1.5.2
joblib==1.4.2
mysql-connector-python==9.0.0

